requirements.txt
dash
pandas
plotly
collections
regex
datetime
gunicorn
dnspython
nbconvert